calvin/william refactored style sheets
