+++ Client/components/login/login.scss
@@ -1,0 +1,114 @@
+.loginContainer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-width: 100vw;
+    min-height: 100vh;
+}
+
+.loginBox {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+    font-size: 60px;
+    border: 1px solid rgb(103, 215, 246);
+    min-width: 300px;
+    min-height: 500px;
+    border-radius: 10px;
+    background-color: rgb(190, 226, 236);
+    box-shadow: 5px 5px 3px rgba(255, 219, 219, 0.798);
+}
+
+.loginWrapper{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    position: absolute;
+    min-width: 100vw;
+}
+
+.loginDetails, .signupDetails {
+    display: flex;
+    align-self: center;
+    flex-direction: column;
+    position: absolute;
+    min-width: 400px;
+    min-height: 400px;
+    border: 1px solid rgb(255, 198, 198);
+    box-shadow: 5px 5px 3px rgba(48, 48, 48, 0.788);
+    border-radius: 10px;
+    background-color: rgba(158, 198, 228, 0.986);
+    z-index: 3;
+    padding: 15px;
+}
+.loginInput {
+    min-width: 51%;
+    min-height: 40px;
+    margin-bottom: 20px;
+    font-size: 18px;
+    font-family: 'Readex Pro', sans-serif;
+    border-radius: 10px;
+    border: 1px solid rgb(238, 167, 167);
+    box-shadow: 3px 3px 2px rgba(232, 200, 200, 0.418);
+    margin-left: 5px;
+}
+
+.loginInput:focus{
+    border: 1px solid rgb(255, 238, 238);
+    outline: none;
+}
+
+.loginText {
+    font-size: 15px;
+    font-family: 'khand', sans-serif;
+    font-weight: bold;
+    margin-top: 3px;
+    color: rgb(72, 159, 163);
+}
+
+.loginTextLarge{
+    font-size: 30px;
+    font-family: 'khand', sans-serif;
+    font-weight: bold;
+    margin-bottom: 40px;
+    height: 60px;
+    color: rgb(255, 255, 255);
+}
+
+.loginButtonBox{
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+    min-height: 200px;
+    min-width: 96%;
+    border-top: 1px solid rgb(249, 203, 234);
+}
+
+.buttonLoginPage {
+    min-width: 75%;
+    border: 1px solid rgb(85, 131, 162);
+    background-color: rgb(116, 221, 179);
+    box-shadow: 3px 3px 2px rgba(160, 197, 155, 0.418);
+    padding: 13px;
+    font-size: 20px;
+    border-radius: 10px;
+    font-family: 'Readex Pro', sans-serif;
+    font-weight: bold;
+    color: rgb(220, 220, 220)
+}
+
+.buttonLoginPage:active{
+    /* rgb(130, 180, 241) */
+    background-color: rgb(116, 177, 251);
+}
+
+.loginFailed{
+    color: rgb(255, 88, 88);
+    font-family: 'Readex Pro', sans-serif;
+    font-weight: bold;
+    font-size: 16px;
+    height: 20px;
+    max-height: 20px;
+}
