+++ Client/components/meal/meal.scss
@@ -1,0 +1,145 @@
+.addMealTextLarge {
+  align-self: flex-start;
+}
+.addMealText {
+  align-self: flex-end;
+}
+.addMealForm {
+  justify-self: flex-end;
+}
+.addMealInput {
+  width: 20vw;
+  margin-right: 20px;
+}
+.addMealInput:focus {
+  border: 1px solid rgb(255, 238, 238);
+  outline: none;
+}
+.addMealText, .addMealTextList, .todaysMeals, .addMealTextLarge {
+  font-size: 15px;
+  font-family: 'khand', sans-serif;
+  font-weight: bold;
+  margin-top: 3px;
+  color: rgb(72, 159, 163);
+}
+.addMeal, .todaysMeals, .dailyGoals, .weeklyGoals {
+  display: flex;
+  flex-direction: column;
+  /* justify-content: space-around; */
+  font-style:'khand', sans-serif;
+  align-items: center;
+  /* font-size: 60px; */
+  border: 1px solid rgba(185, 163, 245, 0.789);
+  border-radius: 10px;
+  background-color: rgb(154, 208, 212);
+  box-shadow: 5px 5px 3px rgba(157, 250, 206, 0.418);
+  color: rgb(255, 255, 255);
+}
+.addMealDetails {
+  display: flex;
+  align-self: center;
+  flex-direction: column;
+  position: absolute;
+  min-width: 400px;
+  min-height: 400px;
+  border: 1px solid rgb(255, 198, 198);
+  box-shadow: 5px 5px 3px rgba(48, 48, 48, 0.788);
+  border-radius: 10px;
+  background-color: rgba(158, 198, 228, 0.986);
+  z-index: 3;
+  padding: 15px;
+}
+.addMealDetails {
+  min-width: 45vw;
+  min-height: 65vh;
+}
+.addMealWrapper{
+  position: absolute;
+  top: 50vh;
+  left: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  position: absolute;
+  min-width: 100vw;
+}
+.submitMealButton {
+  min-width: 75%;
+  border: 1px solid rgb(85, 131, 162);
+  background-color: rgb(116, 221, 179);
+  box-shadow: 3px 3px 2px rgba(160, 197, 155, 0.418);
+  padding: 13px;
+  font-size: 20px;
+  border-radius: 10px;
+  font-family: 'Readex Pro', sans-serif;
+  font-weight: bold;
+  color: rgb(220, 220, 220)
+}
+.submitMealButton{
+  align-self: flex-end;
+  min-width: 25%;
+}
+.submitMealButton:active{
+  /* rgb(130, 180, 241) */
+  background-color: rgb(116, 177, 251);
+}
+.inputWrapper, .addMealDetailsWrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+  min-height: 300px;
+  max-width: 400px;
+  text-align: center;
+}
+.addMealDetailsWrapper {
+  min-width: 100%;
+  min-height: 60vh;
+}
+.addMealInput {
+  min-width: 51%;
+  min-height: 40px;
+  margin-bottom: 20px;
+  font-size: 18px;
+  font-family: 'Readex Pro', sans-serif;
+  border-radius: 10px;
+  border: 1px solid rgb(238, 167, 167);
+  box-shadow: 3px 3px 2px rgba(232, 200, 200, 0.418);
+  margin-left: 5px;
+}
+.addMealNameDetails{
+  display: flex;
+  align-self: center;
+  flex-direction: column;
+  position: absolute;
+  min-width: 400px;
+  min-height: 200px;
+  border: 1px solid rgb(255, 198, 198);
+  box-shadow: 5px 5px 3px rgba(48, 48, 48, 0.788);
+  border-radius: 10px;
+  background-color: rgba(158, 198, 228, 0.986);
+  z-index: 5;
+  padding: 15px;
+}
+.addMealNameDetailsWrapper {
+  min-width: 100%;
+  min-height: 3vh;
+}
+.ingredientList {
+  height: 500px;
+  max-height: 500px;
+  text-align: left;
+  overflow: scroll;
+  align-self: flex-start;
+  padding-left: 30px;
+}
+.ingredientList::-webkit-scrollbar{
+  display: none;
+}
+.listText{
+  text-align: left;
+  align-self: flex-start;
+  margin-left: 20px;
+  font-size: 20px;
+  color: black;
+}
