+++ package.json
@@ -39,7 +39,7 @@
     "html-webpack-plugin": "^5.5.0",
     "jest-environment-jsdom": "^27.4.6",
     "nodemon": "^2.0.15",
-    "sass": "^1.49.0",
+    "sass": "^1.49.7",
     "sass-loader": "^12.4.0",
     "style-loader": "^3.3.1",
     "supertest": "^6.2.2",
