+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from '../Client/app.jsx';
-import styles from '../Client/styles/styles.css';
+import styles from '../Client/styles/styles.scss';
 import logoImg from '../build/image.png'
 
 ReactDOM.render(
