+++ webpack.config.js
@@ -32,8 +32,8 @@ module.exports = {
         }
       },
       {
-        test: /\.css$/i,
-        use: ["style-loader", "css-loader"],
+        test: /\.s?css$/i,
+        use: ["style-loader", "css-loader", "sass-loader"],
       },
       {
         test: /\.png$/i,
